=Editor enhancements. ~Revamped the Great Detective Pikachu previewer with BCFNT support and greatly improved accuracy. ~It is now possible to press Ctrl + A in the textboxes. ~The entry Add, Rename, and Delete functions are now only enabled when the entry list has focus. ~You can now save the generated preview as a PNG. ~Patched the BCFNT class to prevent out of bounds indexes.

+++ Kuriimu/Editor.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using System.Diagnostics;
 using System.Drawing;
+using System.Drawing.Imaging;
 using System.Drawing.Text;
 using System.IO;
 using System.Linq;
@@ -154,19 +155,22 @@ namespace Kuriimu
 
 		private void addEntryToolStripMenuItem_Click(object sender, EventArgs e)
 		{
-			IEntry entry = _fileAdapter.NewEntry();
-
-			frmName name = new frmName(entry, _fileAdapter.EntriesHaveUniqueNames, _fileAdapter.NameList, _fileAdapter.NameFilter, _fileAdapter.NameMaxLength, true);
-
-			if (name.ShowDialog() == DialogResult.OK && name.NameChanged)
+			if (treEntries.Focused)
 			{
-				entry.Name = name.NewName;
-				if (_fileAdapter.AddEntry(entry))
+				IEntry entry = _fileAdapter.NewEntry();
+
+				frmName name = new frmName(entry, _fileAdapter.EntriesHaveUniqueNames, _fileAdapter.NameList, _fileAdapter.NameFilter, _fileAdapter.NameMaxLength, true);
+
+				if (name.ShowDialog() == DialogResult.OK && name.NameChanged)
 				{
-					_hasChanges = true;
-					LoadEntries();
-					treEntries.SelectNodeByIEntry(entry);
-					UpdateForm();
+					entry.Name = name.NewName;
+					if (_fileAdapter.AddEntry(entry))
+					{
+						_hasChanges = true;
+						LoadEntries();
+						treEntries.SelectNodeByIEntry(entry);
+						UpdateForm();
+					}
 				}
 			}
 		}
@@ -177,17 +181,20 @@ namespace Kuriimu
 
 		private void renameEntryToolStripMenuItem_Click(object sender, EventArgs e)
 		{
-			IEntry entry = (IEntry)treEntries.SelectedNode.Tag;
-
-			frmName name = new frmName(entry, _fileAdapter.EntriesHaveUniqueNames, _fileAdapter.NameList, _fileAdapter.NameFilter, _fileAdapter.NameMaxLength);
-
-			if (name.ShowDialog() == DialogResult.OK && name.NameChanged)
+			if (treEntries.Focused)
 			{
-				if (_fileAdapter.RenameEntry(entry, name.NewName))
+				IEntry entry = (IEntry)treEntries.SelectedNode.Tag;
+
+				frmName name = new frmName(entry, _fileAdapter.EntriesHaveUniqueNames, _fileAdapter.NameList, _fileAdapter.NameFilter, _fileAdapter.NameMaxLength);
+
+				if (name.ShowDialog() == DialogResult.OK && name.NameChanged)
 				{
-					_hasChanges = true;
-					treEntries.FindNodeByIEntry(entry).Text = name.NewName;
-					UpdateForm();
+					if (_fileAdapter.RenameEntry(entry, name.NewName))
+					{
+						_hasChanges = true;
+						treEntries.FindNodeByIEntry(entry).Text = name.NewName;
+						UpdateForm();
+					}
 				}
 			}
 		}
@@ -198,17 +205,20 @@ namespace Kuriimu
 
 		private void deleteEntryToolStripMenuItem_Click(object sender, EventArgs e)
 		{
-			IEntry entry = (IEntry)treEntries.SelectedNode.Tag;
-
-			if (MessageBox.Show("Are you sure you want to delete " + entry.Name + "?", "Are you sure?", MessageBoxButtons.YesNo, MessageBoxIcon.Question) == DialogResult.Yes)
+			if (treEntries.Focused)
 			{
-				if (_fileAdapter.DeleteEntry(entry))
+				IEntry entry = (IEntry)treEntries.SelectedNode.Tag;
+
+				if (MessageBox.Show("Are you sure you want to delete " + entry.Name + "?", "Are you sure?", MessageBoxButtons.YesNo, MessageBoxIcon.Question) == DialogResult.Yes)
 				{
-					_hasChanges = true;
-					TreeNode nextNode = treEntries.SelectedNode.NextNode;
-					UpdateEntries();
-					treEntries.Nodes.Remove(treEntries.FindNodeByIEntry(entry));
-					treEntries.SelectedNode = nextNode;
+					if (_fileAdapter.DeleteEntry(entry))
+					{
+						_hasChanges = true;
+						TreeNode nextNode = treEntries.SelectedNode.NextNode;
+						UpdateEntries();
+						treEntries.Nodes.Remove(treEntries.FindNodeByIEntry(entry));
+						treEntries.SelectedNode = nextNode;
+					}
 				}
 			}
 		}
@@ -287,6 +297,22 @@ namespace Kuriimu
 			Settings.Default.Save();
 			UpdatePreview();
 			UpdateForm();
+		}
+
+		private void tsbPreviewSave_Click(object sender, EventArgs e)
+		{
+			SaveFileDialog sfd = new SaveFileDialog();
+			sfd.Title = "Save Preview as PNG...";
+			sfd.InitialDirectory = Settings.Default.LastDirectory;
+			sfd.FileName = "preview.png";
+			sfd.Filter = "Portable Network Graphics (*.png)|*.png";
+			sfd.AddExtension = true;
+
+			if (sfd.ShowDialog() == DialogResult.OK)
+			{
+				Bitmap bmp = (Bitmap)pbxPreview.Image;
+				bmp.Save(sfd.FileName, ImageFormat.Png);
+			}
 		}
 
 		// File Handling
@@ -603,12 +629,12 @@ namespace Kuriimu
 				scbFontSize.Enabled = _fileOpen;
 
 				// Toolbar
-				addEntryToolStripMenuItem.Enabled = canAdd;
-				tsbEntryAdd.Enabled = canAdd;
-				renameEntryToolStripMenuItem.Enabled = canRename;
-				tsbEntryRename.Enabled = canRename;
-				deleteEntryToolStripMenuItem.Enabled = canDelete;
-				tsbEntryDelete.Enabled = canDelete;
+				addEntryToolStripMenuItem.Enabled = canAdd && treEntries.Focused;
+				tsbEntryAdd.Enabled = canAdd && treEntries.Focused;
+				renameEntryToolStripMenuItem.Enabled = canRename && treEntries.Focused;
+				tsbEntryRename.Enabled = canRename && treEntries.Focused;
+				deleteEntryToolStripMenuItem.Enabled = canDelete && treEntries.Focused;
+				tsbEntryDelete.Enabled = canDelete && treEntries.Focused;
 				entryPropertiesToolStripMenuItem.Enabled = itemSelected && _fileAdapter.EntriesHaveExtendedProperties;
 				tsbEntryProperties.Enabled = itemSelected && _fileAdapter.EntriesHaveExtendedProperties;
 				sortEntriesToolStripMenuItem.Enabled = _fileOpen && _fileAdapter.CanSortEntries;
@@ -618,6 +644,7 @@ namespace Kuriimu
 				tsbPreviewEnabled.Enabled = _gameHandler != null ? _gameHandler.HandlerCanGeneratePreviews : false;
 				tsbPreviewEnabled.Image = Settings.Default.PreviewEnabled ? Resources.menu_preview_visible : Resources.menu_preview_invisible;
 				tsbPreviewEnabled.Text = Settings.Default.PreviewEnabled ? "Disable Preview" : "Enable Preview";
+				tsbPreviewSave.Enabled = Settings.Default.PreviewEnabled;
 
 				treEntries.Enabled = _fileOpen;
 
@@ -670,6 +697,16 @@ namespace Kuriimu
 		private void treEntries_AfterCollapse(object sender, TreeViewEventArgs e)
 		{
 			e.Node.Expand();
+		}
+
+		private void treEntries_Enter(object sender, EventArgs e)
+		{
+			UpdateForm();
+		}
+
+		private void treEntries_Leave(object sender, EventArgs e)
+		{
+			UpdateForm();
 		}
 
 		// Text
